tests/dialer/test_incoming_intent.html

<!DOCTYPE HTML>
<html>
  <head>
    <title>Dialer incoming intent</title>
    <script type="text/javascript" src="/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" type="text/css" href="/SimpleTest/test.css" />
  </head>
  <body>
    <p id="display"></p>
    <div id="content" style="display: none">
      <iframe src="/tests/gaia/apps/dialer/dialer.html?choice=incoming"></iframe>
    </div>
    <pre id="test">
    <script class="testbody" type="text/javascript">
      var dialerFrame = document.getElementsByTagName('iframe')[0];

      dialerFrame.addEventListener('load', function runWhenLoaded(evt) {
        dialerFrame.removeEventListener('load', runWhenLoaded);

        var dialerWindow = dialerFrame.contentWindow;
        // faking an incoming call
        dialerWindow.navigator.mozTelephony = {liveCalls: [{number: '123'}]}
        dialerWindow.visibilityChanged(dialerWindow.location.href);

        ok(dialerWindow.document.getElementById('call-screen').classList.contains('oncall'), 'Call screen displayed');
        ok(dialerWindow.document.getElementById('views').classList.contains('oncall'), 'Views in call mode');
        ok(dialerWindow.document.getElementById('tabs-container').classList.contains('oncall'), 'Tabs hidden');
      });
    </script>
    </pre>
  </body>
</html>
